add group tag to opensensemap birdiary

+++ api/api.py
@@ -45,6 +45,7 @@ pwd = os.getenv('Mail_PWD', "ABC")
 
 sensemapUser = os.getenv('opensensemap_User')
 sensemapPwd = os.getenv('opensensemap_PWD')
+API_KEY = os.getenv('API_KEY', "ABC")
 
 
 app = Flask(__name__)
@@ -259,7 +260,7 @@ def index():
 
 
 
-@app.route('/api/image', methods=['Get', 'POST'])
+@app.route('/api/image', methods=['POST'])
 def image():
     if request.method=="POST":
         data = request.files['image']
@@ -270,7 +271,6 @@ def image():
         return jsonify(
             result=result
     )
-    return render_template('./index.html')
 
 @app.route('/api/video', methods=['Get', 'POST'])
 def video():
@@ -360,7 +360,7 @@ def add_station():
                 header = {'content-type': 'application/json', 'Authorization': 'Bearer ' + response['token']}
                 urlOpensensemap = 'https://api.opensensemap.org/boxes/'
                 sensors = [{'title': 'Temperature', 'unit': '°C', 'sensorType': 'DHT22'}, {'title': 'Humidity', 'unit': '%', 'sensorType': 'DHT22'}]
-                data = {'name': name, 'exposure': 'outdoor', 'location': [location.lng, location.lat], 'sensors': sensors} 
+                data = {'name': name, 'exposure': 'outdoor', 'grouptag': 'Birdiary', 'location': [location.lng, location.lat], 'sensors': sensors} 
                 sensemapRequest = requests.post(urlOpensensemap, json=data, headers=header)
 
                 # if sensebox was created successfully add sensebox id
@@ -382,11 +382,21 @@ def station(station_id: str):
             station.measurements.movements = station.measurements.movements[:int(movements)]
         return jsonify(station), 200
     if request.method=="PUT":
+        apikey= request.args.get("apikey")
         body = request.get_json()
+        print(apikey, flush=True)
+        print(API_KEY, flush=True)
+        if API_KEY != apikey:
+            return "Not authorized", 401
         station = Station.objects.get_or_404(station_id=station_id)
         station.update(**body)
         return jsonify(station), 200
     if request.method=="DELETE":
+        apikey= request.args.get("apikey")
+        print(apikey, flush=True)
+        print(API_KEY, flush=True)
+        if API_KEY != apikey:
+            return "Not authorized", 401
         station = Station.objects.get_or_404(station_id=station_id)
         station.delete()
         return jsonify(str(station.station_id)), 200
